feat: add authentications

+++ backend/src/services/posts.js
@@ -1,7 +1,8 @@
 import { Post } from "../db/models/post.js";
+import { User } from "../db/models/user.js";
 
-export async function createPost({ title, author, contents, tags }) {
-  const post = new Post({ title, author, content: contents, tags });
+export async function createPost(userId, { title, contents, tags }) {
+  const post = new Post({ title, author: userId, content: contents, tags });
   return await post.save();
 }
 
@@ -16,8 +17,10 @@ export async function listAllPosts(options) {
   return await listPosts({}, options);
 }
 
-export async function listPostsByAuthor(author, options) {
-  return await listPosts({ author }, options);
+export async function listPostsByAuthor(authorUsername, options) {
+  const user = await User.findOne({ username: authorUsername });
+  if (!user) return [];
+  return await listPosts({ author: user._id }, options);
 }
 
 export async function listPostsByTag(tags, options) {
@@ -28,14 +31,24 @@ export async function getPostById(postId) {
   return await Post.findById(postId);
 }
 
-export async function updatePost(postId, { title, author, contents, tags }) {
+export async function updatePost(
+  userId,
+  postId,
+  { title, author, contents, tags } = {},
+) {
+  const updateData = {};
+  if (title !== undefined) updateData.title = title;
+  if (author !== undefined) updateData.author = author;
+  if (contents !== undefined) updateData.content = contents;
+  if (tags !== undefined) updateData.tags = tags;
+
   return await Post.findOneAndUpdate(
-    { _id: postId },
-    { $set: { title, author, content: contents, tags } },
+    { _id: postId, author: userId },
+    { $set: updateData },
     { new: true },
   );
 }
 
-export async function deletePost(postId) {
-  return await Post.deleteOne({ _id: postId });
+export async function deletePost(userId, postId) {
+  return await Post.deleteOne({ _id: postId, author: userId });
 }
